refactor(backend): rename payload-worker* → payload-documents-worker*
The package was always specifically about the documents collection (it
parses uploads via LlamaParse and writes parsed_text back). The original
name suggested a generic worker. Renames both backend members and their
Python modules to match what they do, and fixes a leak where the shared
`backend/.env` `APP_NAME` was bleeding into the worker's `Settings`
(now hard-coded in `main.py` so each consumer keeps a stable identity).

Touched:

* `backend/payload-worker-builder/` → `payload-documents-worker-builder/`
* `backend/payload-worker/` → `payload-documents-worker/`
* All Python module imports + workspace pyproject members + sources.
* Plugin docstring references updated for symmetry.
