+++ .vscode/launch.json
@@ -26,14 +26,14 @@
       "name": "Worker (kicker)",
       "type": "node-terminal",
       "request": "launch",
-      "command": "uv sync --all-packages && uv run --package payload-worker uvicorn payload_worker.main:app --reload --host 0.0.0.0 --port 8001",
+      "command": "uv sync --all-packages && uv run --package payload-documents-worker uvicorn payload_documents_worker.main:app --reload --host 0.0.0.0 --port 8001",
       "cwd": "${workspaceFolder}/backend"
     },
     {
       "name": "Worker (taskiq)",
       "type": "node-terminal",
       "request": "launch",
-      "command": "uv sync --all-packages && uv run --package payload-worker taskiq worker payload_worker.main:broker --reload",
+      "command": "uv sync --all-packages && uv run --package payload-documents-worker taskiq worker payload_documents_worker.main:broker --reload",
       "cwd": "${workspaceFolder}/backend"
     },
     {
