+++ backend/payload-documents-worker-builder/payload_documents_worker_builder/__init__.py
@@ -1,13 +1,13 @@
-"""Public API for `payload-worker-builder`.
+"""Public API for `payload-documents-worker-builder`.
 
 Mirrors the shape of `agno_agent_builder.__init__`: a single import surface
 that exposes the factory, the config, the clients and the tasks. Consumers
 should never reach into submodules.
 """
 
-from payload_worker_builder.app import WorkerApp, create_app
-from payload_worker_builder.broker import create_broker
-from payload_worker_builder.clients import (
+from payload_documents_worker_builder.app import WorkerApp, create_app
+from payload_documents_worker_builder.broker import create_broker
+from payload_documents_worker_builder.clients import (
     LlamaParseClient,
     LlamaParseError,
     LlamaParseJob,
@@ -15,8 +15,8 @@ from payload_worker_builder.clients import (
     PayloadClient,
     PayloadError,
 )
-from payload_worker_builder.config import RuntimeConfig
-from payload_worker_builder.tasks import (
+from payload_documents_worker_builder.config import RuntimeConfig
+from payload_documents_worker_builder.tasks import (
     PARSE_DOCUMENT_TASK_NAME,
     register_parse_document_task,
 )
