+++ backend/payload-documents-worker-builder/payload_documents_worker_builder/clients/__init__.py
@@ -1,12 +1,12 @@
 """HTTP clients used by built-in tasks (Payload + LlamaParse)."""
 
-from payload_worker_builder.clients.llama_parse import (
+from payload_documents_worker_builder.clients.llama_parse import (
     LlamaParseClient,
     LlamaParseError,
     LlamaParseJob,
     LlamaParseStatus,
 )
-from payload_worker_builder.clients.payload import PayloadClient, PayloadError
+from payload_documents_worker_builder.clients.payload import PayloadClient, PayloadError
 
 __all__ = [
     "LlamaParseClient",
