+++ backend/payload-documents-worker/payload_documents_worker/__init__.py
@@ -1,0 +1,1 @@
+"""PayloadAgents default worker — thin consumer of payload-documents-worker-builder."""
