--- backend/payload-worker/payload_worker/__init__.py
@@ -1,1 +1,0 @@
-"""PayloadAgents default worker — thin consumer of payload-worker-builder."""
